fix _callOnLoss&_spawnLogErr

+++ fed_kit_client/lib/cifar10_ml_client.dart
@@ -78,6 +78,6 @@ class Cifar10MLClient extends MLClient {
   }
 
   void _callOnLoss(loss) {
-    _onLoss?.call(loss);
+    _onLoss?.call(loss.cast<double>());
   }
 }
